updated views with all new functions

+++ slots_game/slots/views.py
@@ -3,19 +3,35 @@ from django.contrib.auth.decorators import login_required, permission_required
 from django.contrib.auth import login, logout, authenticate
 from django.contrib.auth.models import User, Group
 from django.contrib import messages
-import uuid, secrets
-from .models import OneTimeLinkModel
+import uuid, secrets, json
+from .models import OneTimeLinkModel, Game_Settings, Counter, WinDiscription, LobbyDiscription
 
 
-def slots_web(request, access_code):
-    if OneTimeLinkModel.objects.filter(onetime_url=access_code).exists():
-        OneTimeLinkModel.objects.filter(onetime_url=access_code).delete()
-        return render(request, "../templates/slots_template/slots.html", {})
+def game_name_validator(game_name):
+    prizes = list(Game_Settings.objects.values_list('prize_name', flat=True))
+    if game_name in prizes:
+        return True
     else:
-        return render(request, "../templates/admin_signin/darkpan/404.html", {})
+        return False
+
+
+def slots_web(request, access_code, game_name):
+    prizes = Game_Settings.objects.values_list('prize_name', flat=True)
+    data = WinDiscription.objects.values()
+    lobby_data = LobbyDiscription.objects.values()
+    if game_name_validator(game_name):
+        if OneTimeLinkModel.objects.filter(onetime_url=access_code).exists():
+            OneTimeLinkModel.objects.filter(onetime_url=access_code).delete()
+            for game_headers in data:
+                if game_headers['for_game'] == game_name:
+                    return render(request, "../templates/slots_template/slots.html", {'prize': game_headers['prize_name'], 'promocode': game_headers['promocode'], 'company': game_headers['company'], 'link': game_headers['link']})
+        else:
+            return render(request, "../templates/slots_template/lobby.html", {'prizes': list(prizes), 'game_data': json.dumps(list(lobby_data))})
+    else:
+            return render(request, "../templates/slots_template/lobby.html", {'prizes': list(prizes), 'game_data': json.dumps(list(lobby_data))})
     
 
-def slots_game(request, access_code):
+def slots_game(request, access_code, game_name):
     slot_value = secrets.randbelow(10)
     if request.method == 'GET':
         return HttpResponse(slot_value)
@@ -25,7 +41,15 @@ def slots_game(request, access_code):
 
 @login_required(login_url="/login")
 def admin_page(request):
-    return render(request, "../templates/admin_signin/darkpan/index.html", {'username': username})
+    prizes = Game_Settings.objects.values_list('prize_name', flat=True)
+    count_wins = Counter.objects.values_list('count_of_wins', flat=True)
+    count_games = Counter.objects.values_list('count_of_games', flat=True)
+    count_wins = list(count_wins)
+    count_games = list(count_games)
+    try:
+        return render(request, "../templates/admin_signin/darkpan/index.html", {'username': username, 'prizes': list(prizes), 'count_wins': count_wins[-1], 'count_games': count_games[-1]})
+    except:
+        return redirect('login')
 
 def sign_in(request):
     global username
@@ -49,3 +73,16 @@ def generate_url(request):
         OneTimeLinkModel.objects.create(onetime_url=onetime_url)
         print(str(onetime_url))
         return HttpResponse(str(onetime_url))
+
+def counter_data(request, access_code, game_name):
+    if request.method == 'POST':
+        data = json.loads(request.body.decode('utf-8'))['message']
+        if data == 'game':
+            count_games = Counter.objects.last()
+            count_games.count_of_games += 1
+            count_games.save()
+        elif data == 'win':
+            count_wins = Counter.objects.last()
+            count_wins.count_of_wins += 1
+            count_wins.save()
+        return HttpResponse("success")
